fix: updated the readme file
